Compiling on windows

+++ include/hermes_shm/memory/allocator/scalable_page_allocator.h
@@ -195,7 +195,7 @@ class _ScalablePageAllocator : public Allocator {
    * */
   HSHM_CROSS_FUN
   size_t GetCurrentlyAllocatedSize() {
-    return header_->GetCurrentlyAllocatedSize();
+    return (size_t)header_->GetCurrentlyAllocatedSize();
   }
 
   /**
