+++ include/hermes_shm/memory/allocator/stack_allocator.h
@@ -172,7 +172,7 @@ class _StackAllocator : public Allocator {
    * */
   HSHM_CROSS_FUN
   size_t GetCurrentlyAllocatedSize() {
-    return header_->GetCurrentlyAllocatedSize();
+    return (size_t)header_->GetCurrentlyAllocatedSize();
   }
 
   /**
