+++ include/hermes_shm/types/real_number.h
@@ -17,6 +17,7 @@
 #include <cstdint>
 
 #include "hermes_shm/constants/macros.h"
+#include "hermes_shm/types/numbers.h"
 
 #ifdef __cplusplus
 
@@ -44,7 +45,7 @@ struct RealNumber {
   explicit RealNumber(uint64_t numerator, uint64_t denominator) {
     decimal_ = numerator / denominator;
     uint64_t rem = numerator % denominator;
-    numerator_ = rem * precision / denominator;
+    numerator_ = (u32)(rem * precision / denominator);
   }
 
   /**
@@ -92,7 +93,7 @@ struct RealNumber {
   }
 
   HSHM_INLINE_CROSS_FUN
-  size_t as_int() const { return decimal_ + numerator_ / precision; }
+  size_t as_int() const { return (size_t)(decimal_ + numerator_ / precision); }
 };
 
 }  // namespace hshm
